chore(backend): add payment service

+++ server/DesignPatterns/Factories/Abstract/IPaymentFactory.cs
@@ -2,6 +2,6 @@ using server.Dtos.Payment;
 
 namespace server.DesignPatterns.Factories.Abstract {
 	public interface IPaymentFactory {
-		Task<InitPaymentResponse> CreatePaymentAsync(string orderId, string orderDescription, int amount);
+		Task<InitPaymentResponse> CreatePaymentAsync(string orderId, string orderDescription, long amount);
 	}
 }
