+++ server/DesignPatterns/Factories/VnPayAllInOnePaymentFactory.cs
@@ -6,7 +6,7 @@ namespace server.DesignPatterns.Factories {
   {
 		private readonly string API_ENDPOINT_URL = "https://sandbox.vnpayment.vn/paymentv2/vpcpay.html";
 
-		public Task<InitPaymentResponse> CreatePaymentAsync(string orderId, string orderDescription, int amount)
+		public Task<InitPaymentResponse> CreatePaymentAsync(string orderId, string orderDescription, long amount)
 		{
 			throw new NotImplementedException();
 		}
